Work around broken parsing in zulu.bash
There are some entries in the Azul archive that the script does not expect, they are for "JDKs" like `zre9.0.0.15-jre9.0.0-macosx_x64.dmg`. This eventually causes a fatal error, which prevents the script from completing and updating files like `all.json`. This fix just ignores lines that include `zre`.